アイコン取得のロジックを分離
#6
